apply dam's review: autoflow must bypass onboarding gate (#190 blocker)
Fresh sim + autoflow=1 was falling into OnboardingFlow with no taps
available, so screenshot/CI automation stalled instead of reaching the
scripted walk. autoflowRounds is already in scope at AppRoot.init time,
so needsOnboarding's initial value now short-circuits when autoflow is
requested.

Also splits the pre-existing (sac-authored) engine/launch-arg resolution
block out of GalleryApp.swift into EngineResolution.swift — pure
relocation, no logic change — since it had grown the file past the
swiftlint file_length threshold and was blocking every commit to the file.
